project changes

+++ CudaProjectPMPH/OrigImpl/updateParamsKernels.cu.h
@@ -39,7 +39,7 @@ __global__ void setParamsKer(REAL numX ,REAL numY, REAL* myX, REAL* myResult)
     unsigned int j = blockIdx.y*blockDim.y+threadIdx.y;
     unsigned int h = blockIdx.z;
     
-    myResult[h * numX * numY + i * numY + j] = max(myX[i]-0.001*h, (REAL)0.0);
+    myResult[h * numX * numY + i * numY + j] = (REAL) max(myX[i]-0.001*h, (REAL)0.0);
 
 }
